Abstract API search into class and separate helpers into single responsibility files Relates #27 Closes #28
