+++ billy/__init__.py
@@ -1,5 +1,5 @@
-from .helpers import google_book_search, parse_search
 from flask import Flask, render_template, request, redirect, url_for
+from .utils import BookList
 
 
 def create_app(test_config=None):
@@ -25,14 +25,11 @@ def create_app(test_config=None):
         else:
             start = int(request.args.get("start"))
 
-        # Get the search value from the request query string
         query = request.args.get('search')
-        # Hit and parse the API
-        api_search = google_book_search(query, start)
-        parsed_text = parse_search(api_search)
+        book_list_parsed = BookList(query, start).parse()
 
-        if len(parsed_text["items"]) > 0:
-            return render_template("search.html", start=start+1, count=start + len(parsed_text["items"]), search=query, data=parsed_text)
+        if len(book_list_parsed["items"]) > 0:
+            return render_template("search.html", start=start+1, count=start + len(book_list_parsed["items"]), search=query, data=book_list_parsed)
         else:
             return redirect(url_for("index"))
